requirements.txt
absl-py==0.9.0
appnope==0.1.0
argon2-cffi==20.1.0
asn1crypto==1.3.0
astor==0.8.1
async-generator==1.10
attrs==19.3.0
Authlib==0.14.1
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.3.0
boto3==1.11.2
botocore==1.14.2
cachetools==4.1.0
cassandra-driver==3.23.0
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.7.1
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
geomet==0.1.2
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ijson==2.6.1
importlib-metadata==1.6.0
ipykernel==5.5.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lxml==4.6.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
more-itertools==8.2.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.4.5
notebook==6.3.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.1
oscrypto==1.2.0
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.3
paramiko==2.7.1
parso==0.5.2
patsy==0.5.1
pexpect==4.7.0
phonenumbers==8.11.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.10.1
prompt-toolkit==3.0.2
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodomex==3.9.7
Pygments==2.5.2
pyhocon==0.3.54
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.17.3
pytest==5.4.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
shap==0.33.0
six==1.13.0
snowflake-connector-python==2.2.4
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.12
sshtunnel==0.1.5
statsmodels==0.12.2
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
wavio==0.0.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==0.82
zipp==3.1.0
zstandard==0.13.0
